fix: ensure integer division for search count in Tavily and Exa APIs

+++ backend/open_webui/retrieval/web/tavily_exa.py
@@ -23,14 +23,14 @@ def search_tavily_and_exa(
     
     def tavily_search():
         try:
-            return search_tavily(tavily_api_key, query, count / 2 + 1, filter_list)
+            return search_tavily(tavily_api_key, query, int(count / 2) + 1, filter_list)
         except Exception as e:
             log.error(f"Tavily search failed: {str(e)}")
             return []
             
     def exa_search():
         try:
-            return search_exa(exa_api_key, query, count / 2, filter_list)
+            return search_exa(exa_api_key, query, int(count / 2), filter_list)
         except Exception as e:
             log.error(f"Exa search failed: {str(e)}")
             return []
